requirements.txt
fastapi
uvicorn
python-dotenv
requests
pillow
python-multipart